Enhance for Jetpack Beta 3.0.0
Jetpack Beta plugin v3.0.0 adds the ability to work with multiple
plugins. This enhances Jurassic Ninja to take advantage of that,
offering to install a branch for each available plugin and then doing
so.

For this to fully work, you'll have to edit the specialops page with the
change included in this PR.